Revert "Add H3 Enhancer"
