--- dotcom-rendering/src/model/enhance-H3s.ts
@@ -1,82 +1,0 @@
-import { JSDOM } from 'jsdom';
-
-const isFalseH3 = (element: CAPIElement): boolean => {
-	if (!element) return false;
-	// Checks if this element is a 'false h3' based on the convention: <p><strong><H3 text</strong></p>
-	if (
-		element._type !== 'model.dotcomrendering.pageElements.TextBlockElement'
-	) {
-		return false;
-	}
-	const frag = JSDOM.fragment(element.html);
-	if (!frag || !frag.firstElementChild) return false;
-	const html = frag.firstElementChild.outerHTML;
-	// The following things must be true for an element to be a faux H3
-	const hasPwrapper = frag.firstElementChild.nodeName === 'P';
-	const containsStrongtags =
-		frag.firstElementChild.outerHTML.includes('<strong>');
-	const doesNotContainLinks =
-		!frag.firstElementChild.outerHTML.includes('<a>');
-	const startStrong = html.startsWith('<p><strong>');
-	const endsStrong = html.endsWith('</strong></p>');
-
-	return (
-		hasPwrapper &&
-		containsStrongtags &&
-		doesNotContainLinks &&
-		startStrong &&
-		endsStrong
-	);
-};
-
-const extractH3 = (element: CAPIElement): string => {
-	// Extract the text based on the convention: <p><strong><H3 text</strong></p>
-	const textElement = element as TextBlockElement;
-	const frag = JSDOM.fragment(textElement.html);
-	if (isFalseH3(element)) {
-		return (
-			frag.firstElementChild?.innerHTML
-				.split('<strong>')
-				.join('')
-				.split('</strong>')
-				.join('') ?? ''
-		);
-	}
-	return '';
-};
-
-const enhance = (elements: CAPIElement[]): CAPIElement[] => {
-	/**
-	 * Why not just add H3s in Composer?
-	 * Truth is, you can't. So to get around this there's a convention that says if
-	 * you insert <p><strong>Faux H3!</strong>,</p> then we replace it with a h3 tag
-	 * instead.
-	 */
-	const withH3s: CAPIElement[] = [];
-	elements.forEach((thisElement) => {
-		if (
-			thisElement._type ===
-				'model.dotcomrendering.pageElements.TextBlockElement' &&
-			isFalseH3(thisElement)
-		) {
-			const h3Text = extractH3(thisElement);
-
-			withH3s.push({
-				...thisElement,
-				html: `<h3>${h3Text}</h3>`,
-			});
-		} else {
-			// Pass through
-			withH3s.push(thisElement);
-		}
-	});
-	return withH3s;
-};
-
-export const enhanceH3s = (blocks: Block[]): Block[] =>
-	blocks.map((block: Block) => {
-		return {
-			...block,
-			elements: enhance(block.elements),
-		};
-	});
