+++ dotcom-rendering/src/model/enhance-numbered-lists.ts
@@ -1,5 +1,51 @@
 import { JSDOM } from 'jsdom';
 
+const isFalseH3 = (element: CAPIElement): boolean => {
+	if (!element) return false;
+	// Checks if this element is a 'false h3' based on the convention: <p><strong><H3 text</strong></p>
+	if (
+		element._type !== 'model.dotcomrendering.pageElements.TextBlockElement'
+	) {
+		return false;
+	}
+	const frag = JSDOM.fragment(element.html);
+	if (!frag || !frag.firstElementChild) return false;
+	const html = frag.firstElementChild.outerHTML;
+	// The following things must be true for an element to be a faux H3
+	const hasPwrapper = frag.firstElementChild.nodeName === 'P';
+	const containsStrongtags =
+		frag.firstElementChild.outerHTML.includes('<strong>');
+	const doesNotContainLinks =
+		!frag.firstElementChild.outerHTML.includes('<a>');
+	const htmlLength = html.length;
+	const startStrong = html.substr(0, 11) === '<p><strong>';
+	const endsStrong = html.substr(htmlLength - 13) === '</strong></p>';
+
+	return (
+		hasPwrapper &&
+		containsStrongtags &&
+		doesNotContainLinks &&
+		startStrong &&
+		endsStrong
+	);
+};
+
+const extractH3 = (element: CAPIElement): string => {
+	// Extract the text based on the convention: <p><strong><H3 text</strong></p>
+	const textElement = element as TextBlockElement;
+	const frag = JSDOM.fragment(textElement.html);
+	if (isFalseH3(element)) {
+		return (
+			frag.firstElementChild?.innerHTML
+				.split('<strong>')
+				.join('')
+				.split('</strong>')
+				.join('') || ''
+		);
+	}
+	return '';
+};
+
 const isStarRating = (element: CAPIElement): boolean => {
 	const isStar = (charactor: string): boolean => {
 		return charactor === '★' || charactor === '☆';
@@ -181,6 +227,54 @@ const removeGlobalH2Styles = (elements: CAPIElement[]): CAPIElement[] => {
 	return withH2StylesIgnored;
 };
 
+const addH3s = (elements: CAPIElement[]): CAPIElement[] => {
+	/**
+	 * Why not just add H3s in Composer?
+	 * Truth is, you can't. So to get around this there's a convention that says if
+	 * you insert <p><strong>Faux H3!</strong>,</p> then we replace it with a h3 tag
+	 * instead.
+	 *
+	 * Note. H3s don't have any styles so we have to add them. In Frontend, they use
+	 * a 'fauxH3' class for this. In DCR we add `globalH3Styles` which was added at
+	 * the same time as this code.
+	 */
+	const withH3s: CAPIElement[] = [];
+	let previousItem: CAPIElement | undefined;
+	elements.forEach((thisElement) => {
+		if (
+			thisElement._type ===
+				'model.dotcomrendering.pageElements.TextBlockElement' &&
+			isFalseH3(thisElement)
+		) {
+			const h3Text = extractH3(thisElement);
+
+			// To avoid having to depend on the ordering of the enhancer (which could easily be a source of bugs)
+			// We determine if previous items are `ItemLinkBlockElement` through type and `isItemLink` functions
+			const isPreviousItemLink =
+				previousItem?._type ===
+					'model.dotcomrendering.pageElements.ItemLinkBlockElement' ||
+				(previousItem && isItemLink(previousItem));
+
+			withH3s.push(
+				{
+					_type: 'model.dotcomrendering.pageElements.DividerBlockElement',
+					size: 'full',
+					spaceAbove: isPreviousItemLink ? 'loose' : 'tight',
+				},
+				{
+					...thisElement,
+					html: `<h3>${h3Text}</h3>`,
+				},
+			);
+		} else {
+			// Pass through
+			withH3s.push(thisElement);
+		}
+		previousItem = thisElement;
+	});
+	return withH3s;
+};
+
 const addItemLinks = (elements: CAPIElement[]): CAPIElement[] => {
 	const withItemLink: CAPIElement[] = [];
 	elements.forEach((thisElement) => {
@@ -253,46 +347,6 @@ const addTitles = (
 	return withTitles;
 };
 
-const addH3Spacing = (elements: CAPIElement[]): CAPIElement[] => {
-	/**
-	 * Note. H3s don't have any styles so we have to add them. In Frontend, they use
-	 * a 'fauxH3' class for this. In DCR we add `globalH3Styles` which was added at
-	 * the same time as this code.
-	 */
-	const withH3Spacings: CAPIElement[] = [];
-	let previousItem: CAPIElement | undefined;
-	elements.forEach((thisElement) => {
-		if (
-			thisElement._type ===
-				'model.dotcomrendering.pageElements.TextBlockElement' &&
-			thisElement.html.startsWith('<h3>')
-		) {
-			// To avoid having to depend on the ordering of the enhancer (which could easily be a source of bugs)
-			// We determine if previous items are `ItemLinkBlockElement` through type and `isItemLink` functions
-			const isPreviousItemLink =
-				previousItem?._type ===
-					'model.dotcomrendering.pageElements.ItemLinkBlockElement' ||
-				(previousItem && isItemLink(previousItem));
-
-			withH3Spacings.push(
-				{
-					_type: 'model.dotcomrendering.pageElements.DividerBlockElement',
-					size: 'full',
-					spaceAbove: isPreviousItemLink ? 'loose' : 'tight',
-				},
-				{
-					...thisElement,
-				},
-			);
-		} else {
-			// Pass through
-			withH3Spacings.push(thisElement);
-		}
-		previousItem = thisElement;
-	});
-	return withH3Spacings;
-};
-
 class Enhancer {
 	elements: CAPIElement[];
 
@@ -308,11 +362,6 @@ class Enhancer {
 		return this;
 	}
 
-	addH3Spacing() {
-		this.elements = addH3Spacing(this.elements);
-		return this;
-	}
-
 	addTitles() {
 		this.elements = addTitles(this.elements, this.format);
 		return this;
@@ -320,6 +369,11 @@ class Enhancer {
 
 	removeGlobalH2Styles() {
 		this.elements = removeGlobalH2Styles(this.elements);
+		return this;
+	}
+
+	addH3s() {
+		this.elements = addH3s(this.elements);
 		return this;
 	}
 
@@ -347,8 +401,8 @@ const enhance = (
 		new Enhancer(elements, format)
 			// Add the data-ignore='global-h2-styling' attribute
 			.removeGlobalH2Styles()
-			// Add styling to h3 tags
-			.addH3Spacing()
+			// Turn false h3s into real ones
+			.addH3s()
 			// Add item links
 			.addItemLinks()
 			// Add numbered titles
