+++ dotcom-rendering/src/model/enhanceBlocks.ts
@@ -3,7 +3,6 @@ import { enhanceDividers } from './enhance-dividers';
 import { enhanceDots } from './enhance-dots';
 import { enhanceEmbeds } from './enhance-embeds';
 import { enhanceH2s } from './enhance-H2s';
-import { enhanceH3s } from './enhance-H3s';
 import { enhanceImages } from './enhance-images';
 import { enhanceInteractiveContentsElements } from './enhance-interactive-contents-elements';
 import { enhanceNumberedLists } from './enhance-numbered-lists';
@@ -51,11 +50,6 @@ class BlockEnhancer {
 		return this;
 	}
 
-	enhanceH3s() {
-		this.blocks = enhanceH3s(this.blocks);
-		return this;
-	}
-
 	enhanceInteractiveContentsElements() {
 		this.blocks = enhanceInteractiveContentsElements(this.blocks);
 		return this;
@@ -97,7 +91,6 @@ export const enhanceBlocks = (
 ): Block[] => {
 	return new BlockEnhancer(blocks, format, promotedNewsletter)
 		.enhanceDividers()
-		.enhanceH3s()
 		.enhanceH2s()
 		.enhanceInteractiveContentsElements()
 		.enhanceBlockquotes()
